Move scheduler health unit test to utils folder
