+++ tests/cli/commands/test_scheduler_command.py
@@ -17,17 +17,15 @@
 # under the License.
 from __future__ import annotations
 
-from http.server import BaseHTTPRequestHandler
 from importlib import reload
 from unittest import mock
-from unittest.mock import MagicMock
 
 import pytest
 
 from airflow.cli import cli_parser
 from airflow.cli.commands import scheduler_command
 from airflow.executors import executor_loader
-from airflow.utils.scheduler_health import HealthServer, serve_health_check
+from airflow.utils.scheduler_health import serve_health_check
 from airflow.utils.serve_logs import serve_logs
 
 from tests_common.test_utils.config import conf_vars
@@ -224,51 +222,3 @@ class TestSchedulerCommand:
         )
         mock_process.assert_called_once_with(target=serve_logs)
         mock_process().terminate.assert_called_once_with()
-
-
-# Creating MockServer subclass of the HealthServer handler so that we can test the do_GET logic
-class MockServer(HealthServer):
-    def __init__(self):
-        # Overriding so we don't need to initialize with BaseHTTPRequestHandler.__init__ params
-        pass
-
-    def do_GET(self, path):
-        self.path = path
-        super().do_GET()
-
-
-class TestSchedulerHealthServer:
-    def setup_method(self) -> None:
-        self.mock_server = MockServer()
-
-    @mock.patch.object(BaseHTTPRequestHandler, "send_error")
-    def test_incorrect_endpoint(self, mock_send_error):
-        self.mock_server.do_GET("/incorrect")
-        mock_send_error.assert_called_with(404)
-
-    @mock.patch.object(BaseHTTPRequestHandler, "end_headers")
-    @mock.patch.object(BaseHTTPRequestHandler, "send_response")
-    @mock.patch("airflow.utils.scheduler_health.create_session")
-    def test_healthy_scheduler(self, mock_session, mock_send_response, mock_end_headers):
-        mock_scheduler_job = MagicMock()
-        mock_scheduler_job.is_alive.return_value = True
-        mock_session.return_value.__enter__.return_value.query.return_value = mock_scheduler_job
-        self.mock_server.do_GET("/health")
-        mock_send_response.assert_called_once_with(200)
-        mock_end_headers.assert_called_once()
-
-    @mock.patch.object(BaseHTTPRequestHandler, "send_error")
-    @mock.patch("airflow.utils.scheduler_health.create_session")
-    def test_unhealthy_scheduler(self, mock_session, mock_send_error):
-        mock_scheduler_job = MagicMock()
-        mock_scheduler_job.is_alive.return_value = False
-        mock_session.return_value.__enter__.return_value.query.return_value = mock_scheduler_job
-        self.mock_server.do_GET("/health")
-        mock_send_error.assert_called_with(503)
-
-    @mock.patch.object(BaseHTTPRequestHandler, "send_error")
-    @mock.patch("airflow.utils.scheduler_health.create_session")
-    def test_missing_scheduler(self, mock_session, mock_send_error):
-        mock_session.return_value.__enter__.return_value.query.return_value = None
-        self.mock_server.do_GET("/health")
-        mock_send_error.assert_called_with(503)
